Make sure DB schedulers are not decremented if they were never incremented
Previously, accounting for BGWs assigned to DB schedulers was incorrect, because NULL scheduler handlers were translating to BGW_STOPPED. This caused the accounter to decrement BGW_counter despite the counter never being incremented. We fix this by adding an explicit incremented boolean to each scheduler in the hash table. Also did some cleanup with respect to the message-handling functions.