Only load tsl after main loading is done, disable extra warning
We want the deafult license to be Community. Unfortunately, if we load
the tsl during _PG_init parallel workers try to load the tsl before
timescale itself is loaded. To prevent this we defer loading until the
after the main module is loaded.

Also removes the tsl loading warning that fired during normal operation.

Fixes ApacheOnly builds.