Allow downgrade to ApacheOnly license at boot
We do not wish to allow users to downgrade to an ApacheOnly while the
database is running, as this would create an unusual scenario where the
TSL is linked into the running binary but not allowed to be used.
However, now that we have ts_post_load_init, there is a window at
database start when we can downgrade the license before the TSL is
loaded. This commit enables downgrading to an ApacheOnly license iff
the TSL was never loaded in the first place; the intended use is that
the user sets the license to ApacheOnly in their postgresql.conf.

NOTE: downgrading currently can have unpredictable effects.