Improve time_bucket_gapfill optimizations
This commit adds the optimizations we do for regular time bucket,
pushing down index conditions and sort orders to the underlying time
column, to time_bucket_gapfill. This is made slightly more invasive by
the fact that time_bucket_gapfill is marked as VOLATILE (to prevent
postgres from moving it in inappropriate ways) so we much manually fix
up orderings ourselves.